server/requirements.txt
asyncio==3.4.3
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.10
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-vision==2.3.1
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.15.8
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
